requirements.txt
Flask
Werkzeug
Flask-SQLAlchemy
torch
torchvision
Pillow
gunicorn
